+++ ignite/templates/message/message.go
@@ -178,7 +178,7 @@ func typesCodecModify(opts *Options) genny.RunFn {
 		}
 
 		// Import
-		content, err := xast.AppendImports(f.String(), xast.WithLastNamedImport("sdk", "github.com/cosmos/cosmos-sdk/types"))
+		content, err := xast.AppendImports(f.String(), xast.WithNamedImport("sdk", "github.com/cosmos/cosmos-sdk/types"))
 		if err != nil {
 			return err
 		}