+++ ignite/templates/typed/singleton/singleton.go
@@ -570,7 +570,7 @@ func typesCodecModify(opts *typed.Options) genny.RunFn {
 		}
 
 		// Import
-		content, err := xast.AppendImports(f.String(), xast.WithLastNamedImport("sdk", "github.com/cosmos/cosmos-sdk/types"))
+		content, err := xast.AppendImports(f.String(), xast.WithNamedImport("sdk", "github.com/cosmos/cosmos-sdk/types"))
 		if err != nil {
 			return err
 		}