+++ changelog.md
@@ -16,6 +16,7 @@
 - [#4613](https://github.com/ignite/cli/pull/4613) Improve and simplify prompting logic by bubbletea.
 - [#4624](https://github.com/ignite/cli/pull/4624) Fix autocli templates for variadics.
 - [#4645](https://github.com/ignite/cli/pull/4645) Refactor the xast.ModifyFunction to improve the readability.
+- [#4582](https://github.com/ignite/cli/issues/4582) xast misplacing comments.
 
 ### Fixes
 
@@ -25,6 +26,7 @@
 - [#4587](https://github.com/ignite/cli/pull/4587) Add missing light clients routes to IBC client keeper.
 - [#4595](https://github.com/ignite/cli/pull/4595) Fix wrong InterfaceRegistry for IBC modules.
 - [#4609](https://github.com/ignite/cli/pull/4609) Add work dir for relayer integration tests.
+- [#4660](https://github.com/ignite/cli/pull/4660) Fix xast test case indentation.
 
 ### Bug Fixes